# Simpler faster (illustration, for fun)