This replaces a 25-minute benchmark with a 19 seconds one, and it drastically improves the performance in some cases. I have not updated the README, and I image you won't want to merge the PR as is, but the performance boost is significant for some cases.

On an Ice Lake server (with 512-bit registers)....

This PR:

```
Benchmark               (testFile)   Mode  Cnt      Score   Error  Units
BenchJDK.jdk         /twitter.json  thrpt         952.397          ops/s
BenchJDK.scalar      /twitter.json  thrpt        2234.691          ops/s
BenchJDK.vector_128  /twitter.json  thrpt       14630.812          ops/s
BenchJDK.vector_256  /twitter.json  thrpt       17697.810          ops/s
BenchJDK.vector_512  /twitter.json  thrpt       31678.039          ops/s
```

In GB/s, it is...

```
jdk                  0.6 GB/s
scalar             1.4 GB/s
vector_128     9 GB/s
vector_256     11 GB/s
vector_512     20 GB/s
```

These are reasonable results !!!

Previous results (same machine): 

```
Benchmark                          (branchPrediction)              (testFile)   Mode  Cnt       Score        Error  Units
BenchJDK.jdk                                      N/A           /twitter.json  thrpt    5    1730.981 ±     10.969  ops/s
BenchJDK.vector_128                               N/A           /twitter.json  thrpt    5    5312.508 ±     59.222  ops/s
BenchJDK.vector_256                               N/A           /twitter.json  thrpt    5    8186.286 ±     47.844  ops/s
BenchJDK.vector_512                               N/A           /twitter.json  thrpt    5   31273.521 ±    227.835  ops/s
```

So, in my tests, I triple the performance for `vector_128` and double it for `vector_256`.

An interesting addition in this PR is a scalar validation function, which allows you to see directly the benefits of the SIMD operations.

This PR also retains just one benchmark file, because, again, I want quick benchmarks as I do not like to wait several minutes for the results. Ultimately, one would want the input file to be a parameter but I could not get it to work using the instructions in the README (the files are seemingly never found). 

The important point here is that the performance can be made better.