@lemire 2024-06-18
On my ARM-based laptop, the scalar trick does not pay (since it is limited to 128-bit).... but you can still see the massive performance gain compared to scalar with vector_128. In that instance, the JDK 'string creation' trick is faster, likely due to the fact that we have some good optimization in the standard library.

```
BenchJDK.jdk         /twitter.json  thrpt        3857.591          ops/s
BenchJDK.scalar      /twitter.json  thrpt        1977.185          ops/s
BenchJDK.vector_128  /twitter.json  thrpt       19250.184          ops/s
BenchJDK.vector_256  /twitter.json  thrpt          42.616          ops/s
BenchJDK.vector_512  /twitter.json  thrpt          26.642          ops/s
```