@lemire 2024-06-18
My results are several times better than the ones you report in the README...

<img width="876" alt="Screenshot 2024-06-17 at 10 03 55 PM" src="https://github.com/AugustNagro/utf8.java/assets/391987/9fe8ab01-0b01-4970-80e1-9df8897d1c2f">

My CPUs are not special though they are recent.... still you should be getting better results....