@AugustNagro 2024-06-19
Hey I'd love to merge this MR.. this whole repo is just for fun, at least until the vector API stabilizes.

> This replaces a 25-minute benchmark with a 19 seconds one

Yeah totally agree it can be shortened up. My machine requires a bit longer it, but I can explore and adjust that in a separate MR.

> An interesting addition in this PR is a scalar validation function, which allows you to see directly the benefits of the SIMD operations.

Awesome, thanks. Those are excellent additions.

> drastically improves the performance in some cases

I ran the the new benchmark with
* `openjdk 22.0.1+8-16`
* i9-10940X (cascade lake, has slow avx-512)
* `@Warmup(iterations = 3, time = 5)`.

And got these results without the code-change in Utf8.java:

```
Benchmark               (testFile)   Mode  Cnt     Score   Error  Units
BenchJDK.jdk         /twitter.json  thrpt       1541.310          ops/s
BenchJDK.scalar      /twitter.json  thrpt       2519.318          ops/s
BenchJDK.vector_128  /twitter.json  thrpt       5802.873          ops/s
BenchJDK.vector_256  /twitter.json  thrpt       7525.890          ops/s
BenchJDK.vector_512  /twitter.json  thrpt       4861.439          ops/s
```

With your changes:

```
Benchmark               (testFile)   Mode  Cnt      Score   Error  Units
BenchJDK.jdk         /twitter.json  thrpt        1519.961          ops/s
BenchJDK.scalar      /twitter.json  thrpt        2704.991          ops/s
BenchJDK.vector_128  /twitter.json  thrpt       15619.757          ops/s
BenchJDK.vector_256  /twitter.json  thrpt       18113.341          ops/s
BenchJDK.vector_512  /twitter.json  thrpt        4739.419          ops/s
```

Epic! If you re-open I'd be happy to approve & merge.