@lemire 2024-06-19
Re-opened as requested.