@AugustNagro 2024-06-19
Merged; I'll update the readme as well.

I used the OneBranchTooMany benchmark to identify an issue where a branch would cause the vector intrinsics to fail (causing them to box). 

https://mail.openjdk.org/pipermail/panama-dev/2020-December/011650.html

https://mail.openjdk.org/pipermail/panama-dev/2020-December/011651.html

It has been fixed in openjdk now, so I'll remove it.